# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ThemeMetrics API
  description: API für Shopify Theme Performance Analyse
  version: 1.0.0
  contact:
    name: ThemeMetrics Support
    url: https://thememetrics.de
servers:
  - url: https://thememetrics.de/api
    description: Production
  - url: http://localhost:3000/api
    description: Development
security:
  - sessionToken: []
  - cookieAuth: []
...
paths:
...
  /performance:
    get:
      tags: [Analysis]
      summary: Letzte Performance Daten
      parameters:
        - name: shop
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Performance Daten
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PerformanceData'

    post:
      tags: [Analysis]
      summary: PageSpeed Test durchführen
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                  example: https://example.myshopify.com
                strategy:
                  type: string
                  enum: [mobile, desktop]
                  default: mobile
                shop:
                  type: string
      responses:
        '200':
          description: PageSpeed Ergebnis
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageSpeedResult'
  /report:
    get:
      tags: [Dashboard]
      summary: Report Daten für PDF
      parameters:
        - name: shop
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Report Daten
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportData'
  /subscription:
    get:
      tags: [Billing]
      summary: Subscription Status
      parameters:
        - name: shop
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Subscription Info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionInfo'
...
components:
  securitySchemes:
    sessionToken:
      type: http
      scheme: bearer
      description: Shopify Session Token
    cookieAuth:
      type: apiKey
      in: cookie
      name: shop_session
  schemas:
...
    PerformanceData:
      type: object
      properties:
        hasData:
          type: boolean
        mobile:
          type: object
          properties:
            performance:
              type: integer
            lcp:
              type: number
            cls:
              type: number
            tbt:
              type: number
            fcp:
              type: number
        analyzedAt:
          type: string
          format: date-time
    PageSpeedResult:
      type: object
      properties:
        scores:
          type: object
          properties:
            performance:
              type: integer
            accessibility:
              type: integer
            bestPractices:
              type: integer
            seo:
              type: integer
        labData:
          type: object
        fieldData:
          type: object
          nullable: true
        analyzedUrl:
          type: string
        strategy:
          type: string
        timestamp:
          type: string
          format: date-time
    ReportData:
      type: object
      properties:
        store:
          type: object
        agency:
          type: object
          nullable: true
        analysis:
          type: object
        sections:
          type: array
        history:
          type: array
        summary:
          type: object
        generatedAt:
          type: string
          format: date-time
    SubscriptionInfo:
      type: object
      properties:
        plan:
          type: string
          enum: [free, starter, pro, agency]
        status:
          type: string
          enum: [active, cancelled, pending, expired]
        features:
          type: object
        usage:
          type: object
...